Refactor CLI and Controller into their own packages
CLI class and JavaFX controllers are in the main package.

This may lead to confusion and poor code organization.

Let's
* keep CLI only classes and JavaFX Controllers into their own packages

This will increase code modularity.
