+++ src/main/java/terminator/Main.java
@@ -12,6 +12,7 @@ import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.Region;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
+import terminator.controller.DialogBox;
 
 /**
  * Main GUI wrapper element for the application.
